backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netops-assistant-backend"
version = "0.1.0"
description = "Backend API for NetOps Assistant"
requires-python = ">=3.14"
dependencies = [
  "fastapi==0.135.3",
  "uvicorn[standard]==0.35.0",
  "pydantic-settings==2.11.0",
  "sqlalchemy[asyncio]==2.0.44",
  "asyncpg==0.30.0",
  "pwdlib[argon2]==0.3.0",
  "ldap3==2.9.1",
  "alembic==1.16.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "pytest-asyncio==1.2.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
